backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pymongo==4.12.1
groq==0.25.0
huggingface-hub==0.30.2
tensorflow-cpu==2.19.0
torch==2.7.0
torchvision==0.22.0
efficientnet-pytorch==0.7.1
Pillow==11.2.1
numpy==2.1.3
python-dotenv==1.1.0
httpx==0.28.1
sse-starlette==2.3.3
bleach==6.2.0
